feat: Add button to show current network
