+++ web/src/components/sidebars/AnalyticsPanel.vue
@@ -20,8 +20,7 @@ import {
 } from "@/api/rest";
 import NodeAnimation from "./NodeAnimation.vue";
 import { AnalysisResult, Layer, Chart } from "@/types";
-import { addLayer } from "@/layers";
-
+import { isVisible, show, showableTypes } from "@/storeFunctions"
 
 const searchText = ref();
 const filteredAnalysisTypes = computed(() => {
@@ -81,124 +80,6 @@ const inputSelectionRules = [
 const additionalAnimationLayers = ref();
 const inputForm = ref();
 const ws = ref();
-
-function isVisible(value: any): boolean {
-  if (value.type == 'Chart') {
-    const chartPanel = panelArrangement.value.find((panel) => panel.id === 'charts')
-    if (!chartPanel) return false;
-    return currentChart.value?.id == value.id && chartPanel.visible
-  } else if (value.type === 'Dataset') {
-    return selectedLayers.value.some((layer) => {
-      return layer.dataset.id === value.id && layer.visible
-    })
-  } else if (value.type === 'Layer') {
-    return selectedLayers.value.some((layer) => {
-      return layer.id === value.id && layer.visible
-    });
-  } else if (value.type === 'Network') {
-    return isVisible({
-      ...value.dataset,
-      type: 'Dataset',
-    })
-  } else if (value.type === 'AnalysisResult') {
-    const analysisType = availableAnalysisTypes.value?.find((t) => t.db_value === value.analysis_type)
-    if (analysisType) {
-      const showables: Record<string, any>[] = []
-       Object.entries(value.outputs).forEach(
-        ([outputKey, outputValue]) => {
-          const type = analysisType?.output_types[outputKey]
-          if (showableTypes.includes(type)) {
-            showables.push({
-              id: outputValue,
-              type
-            })
-          }
-        }
-      );
-      Object.entries(value.inputs).forEach(
-        ([inputKey, inputValue])=> {
-          const type = analysisType?.input_types[inputKey]
-          const value: Record<string, any> = analysisType.input_options[inputKey]?.find((o: any) => o.id === inputValue)
-          if (showableTypes.includes(type)) {
-            showables.push({
-              ...value,
-              type
-            })
-          }
-        }
-      );
-      return showables.every((o) => isVisible(o))
-    }
-  }
-  return false;
-}
-
-function show(value: any) {
-  if (value.type === 'Chart') {
-    const chartPanel = panelArrangement.value.find((panel) => panel.id === 'charts')
-    if (chartPanel && !chartPanel?.visible) chartPanel.visible = true
-    currentChart.value = value as Chart
-  } else if (value.type === 'Dataset') {
-    getDatasetLayers(value.id).then((layers) => {
-      layers.forEach((layer) => {
-        show({
-          ...layer,
-          type: 'Layer'
-        })
-      })
-    })
-  } else if (value.type === 'Layer') {
-    let add = true
-    selectedLayers.value = selectedLayers.value.map((layer) => {
-        if (add && layer.id === value.id) {
-          layer.visible = true;
-          add = false;
-        }
-        return layer
-    })
-    if (add) addLayer(value as Layer)
-  } else if (value.type === 'Network') {
-    show({
-      ...value.dataset,
-      type: 'Dataset',
-    })
-  } else if (value.type === 'AnalysisResult') {
-    const analysisType = availableAnalysisTypes.value?.find((t) => t.db_value === value.analysis_type)
-    if (analysisType) {
-      Object.entries(value.outputs).map(([outputKey, outputValue]) => {
-        const type = analysisType.output_types[outputKey]
-        if (showableTypes.includes(type)) {
-          show({
-            id: outputValue,
-            type
-          })
-        }
-      })
-      Object.entries(value.inputs).map(([inputKey, inputValue]) => {
-        const type = analysisType.input_types[inputKey]
-        const value: Record<string, any> = analysisType.input_options[inputKey].find((o: any) => o.id === inputValue)
-        if (showableTypes.includes(type)) {
-          show({
-            ...value,
-            type
-          })
-        }
-      })
-    }
-  }
-   else if (['RasterData', 'VectorData'].includes(value.type)) {
-    if (value.dataset) {
-      getDataset(value.dataset).then((dataset) => {
-        show({
-          ...dataset,
-          type: 'Dataset'
-        })
-      })
-    }
-   }
-}
-
-const showableTypes = ['Chart', 'Dataset', 'Network', 'Layer', 'AnalysisResult', 'RasterData', 'VectorData']
 
 function run() {
   inputForm.value.validate().then(({ valid }: { valid: boolean }) => {
